requirements.txt
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.32.0
plotly>=5.18.0
ujson>=5.8.0
tqdm>=4.66.0
openpyxl>=3.1.0
google-cloud-bigquery>=3.20.0
db-dtypes>=1.2.0
